pyproject.toml
[project]
name = "dao_python"
version= "0.0.1"
authors = [
    {name = "Torok Mark Patrik", email = "torok.mark.patrik@gmail.com"},
]
description = "Pet project for learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["eth-brownie==1.19.3", "pydantic==1.10.9"]


[build-system]
requires = ["setuptools", "wheel"]

[project.optional-dependencies]
test = [
    "pytest~=7.3.2",
    "pytest-cov~=4.1.0",
]

lint = [
    "black==22.10.0",
    "ruff==0.0.272",
]

dev=["dao_python[test]","dao_python[lint]"]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
